Decouple select option populating and input form
This commits unties the populating of custom select menu options from
the creation of an input form. This enables these select menus to be
used in other areas of the code without having to deal with an input
form. To accomplish this, a global provider system is created along with
a custom HTML select element
