+++ src/common/inputForm.ts
@@ -2,7 +2,7 @@ import "../common/htmlElement"
 import "../common/htmlFormElement"
 import "../common/date"
 import { FormInputElement, simplifyInputName } from "../common/htmlFormElement";
-import { SelectMenuElement } from "./selectMenuElement";
+import { BetterSelectMenu } from "./selectInputs";
 
 function isInputElement(_obj: any): _obj is FormInputElement  {
     return true;
@@ -11,19 +11,16 @@ function isInputElement(_obj: any): _obj is FormInputElement  {
 class InputForm {
     formElement: HTMLFormElement
     inputs: Map<String, FormInputElement>
-    customSelectInputs: Map<String, SelectMenuElement>
-    selectInputOptionsProvider: Record<string, any> 
+    customSelectInputs: Map<String, BetterSelectMenu>
 
     constructor(
         formClass: string, 
         onSubmit: (json: unknown) => void, 
         onReset: (e: Event) => void,
-        selectInputOptionsProvider: Record<string, any> = {}
     ) {
         this.inputs = new Map();
         this.customSelectInputs = new Map();
         this.formElement = <HTMLFormElement>document.getElementsByClassName(formClass)[0];
-        this.selectInputOptionsProvider = selectInputOptionsProvider
 
         this.formElement.addEventListener('submit', e => {
             e.preventDefault();
@@ -47,7 +44,6 @@ class InputForm {
 
         this.formElement.addEventListener('reset', e => {
             e.preventDefault();
-            
             onReset(e);
 
             return false;
@@ -81,9 +77,6 @@ class InputForm {
                     updateValidationMessage()
                 }
             }
-
-            if(e instanceof HTMLSelectElement || e.getAttribute("role") === "combobox")
-                this.initSelectMenu(e, selectInputOptionsProvider);
 
             // Add unit selection dropdowns
             const useUnits = e.getAttribute('use-units')
@@ -129,7 +122,7 @@ class InputForm {
     clear() {
         this.inputs.forEach(input => {
             const type = input.getAttribute('type');
-            if(SelectMenuElement.isCustomSelect(input)) {
+            if(BetterSelectMenu.isCustomSelect(input)) {
                 return;
             }
 
@@ -215,12 +208,12 @@ class InputForm {
                 continue
 
             // 
-            if(SelectMenuElement.isCustomSelect(element)) {
+            if(BetterSelectMenu.isCustomSelect(element)) {
                 if(!element.parentElement?.parentElement)
                     continue;
                 const options = Array.from(element.parentElement.parentElement.getElementsByTagName("li"));
                 const optionId = options.filter(e => e.getAttribute("value")?.endsWith(obj[key]))[0]?.id;
-                SelectMenuElement.setSelectMenuSelectedOption(element as HTMLInputElement, options, optionId)
+                BetterSelectMenu.setSelectMenuSelectedOption(element as HTMLInputElement, options, optionId)
             } else if(element as HTMLInputElement) {
                 if(element.getAttribute("type") === "checkbox" && element)
                     (element as HTMLInputElement).checked = obj[key];
@@ -243,34 +236,6 @@ class InputForm {
 
             this.setChecked(input.id, this.hasValue(toggles))
         })
-    }
-
-    initSelectMenu(element: FormInputElement, selectInputOptionsProvider: Record<string, any> = {}) {
-        const optionsFrom = element.getAttribute("options-from");
-        if(!optionsFrom) {
-            console.error(`Select element \'${element.name}\' does not have a valid \'options-from\` attribute.`);
-            return;
-        }
-    
-        // Convert the corresponding enum type to its keys
-        const enumObj = selectInputOptionsProvider[optionsFrom];
-        const optionStrings = Object.keys(enumObj);
-        if(!optionStrings) {
-            console.error(`Failed to find registered select options provider called \'${optionsFrom}\'`);
-            return;
-        }
-    
-        if(element.getAttribute("role") === "combobox") {
-            this.customSelectInputs.set(element.id, new SelectMenuElement(element as HTMLInputElement, optionStrings, enumObj));
-        } else {
-            element.append(...optionStrings.map(option => {
-                const optionElement = document.createElement("option");
-                optionElement.innerText = enumObj[option]; // Get enum name as string
-                optionElement.setAttribute("value", option);
-    
-                return optionElement;
-            }));
-        }
     }
 }
