+++ src/renderer/index.html
@@ -119,7 +119,7 @@
     </menu>
 
     <script type="module" src="./src/renderer.ts"></script>
-    <!-- <script type="module" src="./src/nav.ts"></script> -->
+    <script type="module" src="./src/home.ts"></script>
     <script type="module" src="./src/reminder.ts"></script>
     <script type="module" src="./src/settings.ts"></script>
 </body>
