+++ src/renderer/reminder.html
@@ -60,7 +60,7 @@
             >
         </div>
 
-        <div class="combobox-wrapper input-container-grid" title="The type of notification sent for the reminder">
+        <better-select-menu class="combobox-wrapper input-container-grid" title="The type of notification sent for the reminder">
             <!-- Accessible select as per the ARIA 1.2 standards: https://www.w3.org/TR/wai-aria-1.2/#combobox -->
             <label for="notification-type">Notification Type</label>
             <div class="combobox--input">
@@ -86,10 +86,10 @@
                 
                 </ul>
             </div>
-        </div>
+        </better-select-menu>
 
         <div class="input-container-flex">
-            <div class="combobox-wrapper" id="reminder-audio-combobox" title="The audio played when the reminder is sent">
+            <better-select-menu class="combobox-wrapper" id="reminder-audio-combobox" title="The audio played when the reminder is sent">
                 <!-- Accessible select as per the ARIA 1.2 standards: https://www.w3.org/TR/wai-aria-1.2/#combobox -->
                 <label for="reminder-audio-id">Reminder Audio</label>
                 <div class="combobox--input">
@@ -116,14 +116,14 @@
                     
                     </ul>
                 </div>
-            </div>
+            </better-select-menu>
             
-            <button class="primary svg-button" id="play-reminder-audio" type="button" title="Play the selected audio">
+            <button class="primary svg-button play-audio" audioSource="reminder-audio-id" type="button" title="Play the selected audio">
                 <svg xmlns="http://www.w3.org/2000/svg" height="24" viewBox="0 -960 960 960" width="24">
                      <path d="M560-131v-82q90-26 145-100t55-168q0-94-55-168T560-749v-82q124 28 202 125.5T840-481q0 127-78 224.5T560-131ZM120-360v-240h160l200-200v640L280-360H120Zm440 40v-322q47 22 73.5 66t26.5 96q0 51-26.5 94.5T560-320Z"/>
                  </svg>
              </button>
-             <button class="primary svg-button" id="upload-reminder-audio" type="button" title="Upload new audio">
+             <button class="primary svg-button upload-reminder-audio" type="button" title="Upload new audio">
                  <svg xmlns="http://www.w3.org/2000/svg" height="24" viewBox="0 -960 960 960" width="24">
                      <path d="M440-440H200v-80h240v-240h80v240h240v80H520v240h-80v-240Z"/>
                  </svg>
